refactor: remove RadioPanel/index.ts to match existing component pattern
Other components in src/components/ don't use index.ts barrel files;
import RadioPanel directly from its module file instead.
