+++ src/applications/SvarSed/EndredeForhold/EndredeForhold.tsx
@@ -1,5 +1,5 @@
 import {Box, Heading, HStack, RadioGroup, VStack} from '@navikt/ds-react'
-import RadioPanel from 'components/RadioPanel'
+import RadioPanel from 'components/RadioPanel/RadioPanel'
 import { resetValidation, setValidation } from 'actions/validation'
 import { validateEndredeForhold, ValidationEndredeForholdProps } from 'applications/SvarSed/EndredeForhold/validation'
 import { MainFormProps, MainFormSelector } from 'applications/SvarSed/MainForm'
