+++ src/applications/SvarSed/PersonOpplysninger/PersonOpplysninger.tsx
@@ -1,5 +1,5 @@
 import {Alert, Box, Heading, HGrid, HStack, RadioGroup, Spacer, TextField, VStack} from '@navikt/ds-react'
-import RadioPanel from 'components/RadioPanel'
+import RadioPanel from 'components/RadioPanel/RadioPanel'
 import { resetValidation, setValidation } from 'actions/validation'
 import { MainFormProps, MainFormSelector } from 'applications/SvarSed/MainForm'
 import {
