+++ src/applications/SvarSed/Ugyldiggjøre/Ugyldiggjøre.tsx
@@ -1,5 +1,5 @@
 import {Box, Heading, HGrid, RadioGroup, VStack} from '@navikt/ds-react'
-import RadioPanel from 'components/RadioPanel'
+import RadioPanel from 'components/RadioPanel/RadioPanel'
 import { resetValidation, setValidation } from 'actions/validation'
 import { MainFormProps, MainFormSelector } from 'applications/SvarSed/MainForm'
 import Input from 'components/Forms/Input'
