+++ src/applications/SvarSed/YtterligereInfoOmKrav/YtterligereInfoOmKrav.tsx
@@ -1,5 +1,5 @@
 import {Box, Heading, HStack, RadioGroup, Spacer, VStack} from '@navikt/ds-react'
-import RadioPanel from 'components/RadioPanel'
+import RadioPanel from 'components/RadioPanel/RadioPanel'
 import { resetValidation, setValidation } from 'actions/validation'
 import { validateYtterligereInfoOmKrav, ValidationYtterligereInfoOmKravProps } from 'applications/SvarSed/YtterligereInfoOmKrav/validation'
 import { MainFormProps, MainFormSelector } from 'applications/SvarSed/MainForm'
