--- src/components/RadioPanel/index.ts
@@ -1,2 +1,0 @@
-export { default } from './RadioPanel'
-export type { RadioPanelProps } from './RadioPanel'
